refactor(dto): extract shared UserResponse into internal/pkg/dto
Removes duplicate UserResponse struct and NewUserResponse function
from auth and profile delivery packages.
