+++ internal/auth/delivery/http/dto.go
@@ -1,11 +1,5 @@
 package http
 
-import (
-	"time"
-
-	"github.com/go-park-mail-ru/2026_1_KISS/internal/domain"
-)
-
 type RegisterRequest struct {
 	Username string `json:"username"`
 	Email    string `json:"email"`
@@ -15,28 +9,4 @@ type RegisterRequest struct {
 type LoginRequest struct {
 	Email    string `json:"email"`
 	Password string `json:"password"`
-}
-
-type UserResponse struct {
-	ID          int64     `json:"id"`
-	Username    string    `json:"username"`
-	Email       string    `json:"email"`
-	AvatarURL   string    `json:"avatar_url"`
-	Status      string    `json:"status"`
-	Description string    `json:"description"`
-	CreatedAt   time.Time `json:"created_at"`
-	UpdatedAt   time.Time `json:"updated_at"`
-}
-
-func NewUserResponse(u *domain.User) UserResponse {
-	return UserResponse{
-		ID:          u.ID,
-		Username:    u.Username,
-		Email:       u.Email,
-		AvatarURL:   u.AvatarURL,
-		Status:      u.Status,
-		Description: u.Description,
-		CreatedAt:   u.CreatedAt,
-		UpdatedAt:   u.UpdatedAt,
-	}
 }
